fix unit tests so they run across all configured test platforms

+++ symmetric-oracle/src/test/java/org/jumpmind/symmetric/io/data/writer/OracleBulkDatabaseWriterTest.java
@@ -27,6 +27,7 @@ import junit.framework.Assert;
 
 import org.jumpmind.db.DbTestUtils;
 import org.jumpmind.db.platform.oracle.OracleDatabasePlatform;
+import org.jumpmind.db.util.BasicDataSourcePropertyConstants;
 import org.jumpmind.symmetric.io.data.CsvData;
 import org.jumpmind.symmetric.io.data.DataEventType;
 import org.junit.Before;
@@ -38,9 +39,13 @@ public class OracleBulkDatabaseWriterTest extends AbstractWriterTest {
 
     @BeforeClass
     public static void setup() throws Exception {
-        platform = DbTestUtils.createDatabasePlatform(DbTestUtils.ROOT);
-        platform.createDatabase(platform.readDatabaseFromXml("/testOracleBulkWriter.xml", true),
-                true, false);
+        if (DbTestUtils.getEnvironmentSpecificProperties(DbTestUtils.ROOT)
+                .get(BasicDataSourcePropertyConstants.DB_POOL_DRIVER)
+                .equals("oracle.jdbc.driver.OracleDriver")) {
+            platform = DbTestUtils.createDatabasePlatform(DbTestUtils.ROOT);
+            platform.createDatabase(
+                    platform.readDatabaseFromXml("/testOracleBulkWriter.xml", true), true, false);
+        }
     }
 
     @Before
@@ -61,7 +66,7 @@ public class OracleBulkDatabaseWriterTest extends AbstractWriterTest {
 
     @Test
     public void testInsert1000Rows() {
-        if (platform instanceof OracleDatabasePlatform) {
+        if (platform != null && platform instanceof OracleDatabasePlatform) {
             platform.getSqlTemplate().update("truncate table test_bulkload_table_1");
 
             List<CsvData> datas = new ArrayList<CsvData>();
@@ -83,7 +88,7 @@ public class OracleBulkDatabaseWriterTest extends AbstractWriterTest {
 
     @Test
     public void testInsertCollision() {
-        if (platform instanceof OracleDatabasePlatform) {
+        if (platform != null && platform instanceof OracleDatabasePlatform) {
             platform.getSqlTemplate().update("truncate table test_bulkload_table_1");
 
             String[] values = { getNextId(), "string2", "string not null2", "char2",
